pyproject.toml
[tool.poetry]
name = "military-grants"
version = "0.1.0"
description = ""
authors = ["Asura Enkhbayar <asura.enkhbayar@gmail.com>"]
packages = [
    { include = "src" },
]

[tool.poetry.dependencies]
python = "^3.6.10"
pandas = "^1.0.1"
python-dotenv = "^0.12.0"
tqdm = "^4.43.0"
click = "^7.0"
xlrd = "^1.2.0"
requests = "^2.23.0"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
flake8 = "^3.7.9"
Sphinx = "^2.4.3"
coverage = "^5.0.3"
ipykernel = "^5.1.4"
ipywidgets = "^7.5.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
